# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Evaluation:
    properties:
      Note:
        type: string
      Rating:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - Rating
    type: object
  dto.BatchUpdateRequest:
    properties:
      increment:
        type: integer
    required:
    - increment
    type: object
  dto.CreateEvaluationRequest:
    properties:
      Note:
        type: string
      Rating:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - Rating
    type: object
  dto.EvaluationResponse:
    properties:
      id:
        type: integer
      note:
        type: string
      rating:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UpdateEvaluationRequest:
    properties:
      Note:
        type: string
      Rating:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        type: integer
    required:
    - Rating
    type: object
host: localhost:8080
info:
  contact: {}
  description: Essa API permite manter todas as avaliações realizadas.
  title: API de avaliações
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o utilizador e gera o token para os próximos acessos
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      summary: Realizar autenticação
  /auth/refresh_token:
    put:
      consumes:
      - application/json
      description: Atualiza o token de autenticação do usuário
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      security:
      - BearerAuth: []
      summary: Atualiza token de autenticação
  /auth/register:
    post:
      consumes:
      - application/json
      description: Regista um utilizador
      parameters:
      - description: Registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
        "401":
          description: Unauthorized
      summary: Realizar registro
  /echo:
    get:
      consumes:
      - application/json
      description: Echo the data sent through the get request.
      parameters:
      - description: string valid
        in: query
        maxLength: 10
        minLength: 1
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
      summary: Echo the data sent on get
  /evaluation:
    get:
      consumes:
      - application/json
      description: Exibe a lista de todas as avaliações
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Evaluation'
            type: array
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Recupera as avaliações
    post:
      consumes:
      - application/json
      description: Cria uma avaliação sobre a utilização da aplicação
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add evaluation
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEvaluationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EvaluationResponse'
        "400":
          description: Bad request
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Adicionar uma avaliação
  /evaluation/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui uma avaliação realizada
      operationId: delete-evaluation-by-int
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Exclui uma avaliação pelo ID
    get:
      consumes:
      - application/json
      description: Exibe os detalhes de uma avaliação pelo ID
      operationId: get-evaluation-by-int
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Evaluation'
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Recupera uma avaliação pelo id
    put:
      consumes:
      - application/json
      description: Atualiza uma avaliação sobre a utilização da aplicação
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update evaluation
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEvaluationRequest'
      - description: Evaluation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EvaluationResponse'
        "400":
          description: Bad request
        "404":
          description: Not found
      security:
      - BearerAuth: []
      summary: Atualiza uma avaliação
  /evaluation/audit:
    post:
      consumes:
      - application/json
      description: Demonstrates manual transaction with rollback/commit control
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add evaluation
        in: body
        name: evaluation
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEvaluationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EvaluationResponse'
        "400":
          description: Bad request
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Create evaluation with audit log using transaction
  /evaluation/batch:
    put:
      consumes:
      - application/json
      description: Demonstrates GORM Transaction handling - batch update
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Increment value
        in: body
        name: increment
        required: true
        schema:
          $ref: '#/definitions/dto.BatchUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Batch update ratings with transaction
  /evaluation/raw:
    get:
      consumes:
      - application/json
      description: Demonstrates GORM Raw SQL query with parameter binding
      parameters:
      - description: Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Minimum rating filter
        in: query
        name: min_rating
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Evaluation'
            type: array
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get evaluations using raw SQL query
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"