# internal/domain/evaluation.go
package domain

import "gorm.io/gorm"

// Evaluation represents a user evaluation/rating in the system
type Evaluation struct {
	gorm.Model `swaggerignore:"true"`
	Rating     int    `json:"Rating" binding:"required,oneof=0 1 2 3 4 5"`
	Note       string `json:"Note"`
}